Refactoring code, moving functions to helper script file
