+++ src/helpers/helpers.ts
@@ -377,7 +377,7 @@ export async function quickPickSelectItemDelete (items:any) {
             if (!ocapiCall.error) {
                 vscode.window.showInformationMessage(Constants.CODEVERSION_SUCCESS_FIRST + selectionItem + Constants.DELETE_ITEM_SUCCESS_SECOND + currentJson.hostname);
             } else {
-                vscode.window.showInformationMessage(Constants.DELETE_ITEM_ERROR + selectionItem);
+                vscode.window.showErrorMessage(Constants.DELETE_ITEM_ERROR + selectionItem);
             }         
             
         })
@@ -397,7 +397,7 @@ export async function inputboxCreateItem () {
     });
 
     if (inputText && inputText.length <= 3) {
-        vscode.window.showInformationMessage(Constants.INPUTBOX_ERROR_LENGTH);
+        vscode.window.showErrorMessage(Constants.INPUTBOX_ERROR_LENGTH);
         return;
     }
     
@@ -408,8 +408,59 @@ export async function inputboxCreateItem () {
         if (!ocapiCall.error) {
             vscode.window.showInformationMessage(Constants.CODEVERSION_SUCCESS_FIRST + inputText + Constants.INPUTBOX_SUCCESS_SECOND + currentJson.hostname);
         } else {
-            vscode.window.showInformationMessage(Constants.INPUTBOX_ERROR + inputText);
+            vscode.window.showErrorMessage(Constants.INPUTBOX_ERROR + inputText);
         }    
     }
     
 }
+
+/**
+ * Open quickpick to select a Code Version or Hostname
+ *
+ * @param   items List of Code Version or Hostname to show on quickpick
+ * @param   title Title of the quickpick
+ * @param   jsonField Name of property of the dw.json file
+ * @param   propertyChange Property of settings to be updated
+ * @param   remoteAccess If is execution to remote environment
+ */
+export async function quickPickSelectItem (items:any, title:string, jsonField:string, propertyChange:any, remoteAccess:boolean) {
+    return new Promise<void>((resolve) => {
+      const quickPick = vscode.window.createQuickPick();
+      quickPick.items = items.map((item: any) => ({ label: item.displayName, id: item.id }));
+      
+      quickPick.title = title;
+                          
+      quickPick.onDidAccept(async () => {
+        //@ts-ignore
+        const selectionText = quickPick.activeItems[0].id;
+        const currentJson:any = defaultJson();
+        const { writeFileSync } = require("fs");
+        const path = jsonPath(); 
+
+        currentJson[jsonField] = selectionText;        
+        writeFileSync(path, JSON.stringify(currentJson, null, 2), "utf8");
+        quickPick.hide();
+        
+        
+        // Only runs in scenario of getting Code Versions on remote environment, will activate the Code Version on environment
+        if (remoteAccess) {
+
+            // On scenario of getting Code Versions on remote environment, will be update the array of CodeversionsHistory 
+            if (propertyChange) {
+                updateProperty(selectionText, propertyChange); 
+            }
+
+            const ocapiCall:any = await ocapiActiveCodeVersion(selectionText);
+
+            if (!ocapiCall.error) {
+                vscode.window.showInformationMessage(Constants.CODEVERSION_SUCCESS_FIRST + selectionText + Constants.ACTIVE_CODEVERSION_SUCCESS_SECOND + currentJson.hostname);
+            }            
+        }   
+        
+        resolve();
+      })
+      
+      quickPick.show();
+    }
+    )
+  }
